+++ backend/dataall/core/stacks/api/resolvers.py
@@ -61,11 +61,8 @@ def resolve_stack_visibility(context, source: Stack, **kwargs):
     if not source:
         return False
     log_config = config.get_property(StackServiceUtils.map_target_to_config(target_type=source.stack), 'enabled')
-    try:
-        return StackServiceUtils.check_if_user_allowed_view_logs(context.groups, log_config)
-    except Exception as e:
-        log.error(f'Failed to check if user is allowed to view logs: {e}')
-        return False
+    return StackServiceUtils.check_if_user_allowed_view_logs(context.groups, log_config)
+
 
 
 def resolve_task_id(context, source: Stack, **kwargs):
