+++ backend/dataall/modules/shares_base/services/share_logs_service.py
@@ -2,9 +2,9 @@ import os
 import logging
 
 from dataall.base.context import get_context
+from dataall.base.feature_toggle_checker import is_feature_enabled_for_allowed_values
 from dataall.base.utils import Parameter
 from dataall.base.db import exceptions
-from dataall.base.utils.logs import is_feature_has_allowed_values
 from dataall.core.stacks.aws.cloudwatch import CloudWatch
 from dataall.base.config import config
 
@@ -16,20 +16,17 @@ log = logging.getLogger(__name__)
 
 class ShareLogsService:
     @staticmethod
-    @is_feature_has_allowed_values(
+    @is_feature_enabled_for_allowed_values(
         allowed_values=['admin-only', 'enabled', 'disabled'],
+        enabled_values=['admin-only', 'enabled'],
         default_value='enabled',
         config_property='modules.shares_base.features.show_share_logs',
     )
     def check_view_logs_permissions(username, groups, shareUri):
         context = get_context()
         log_config = config.get_property('modules.shares_base.features.show_share_logs', 'enabled')
-        if (log_config == 'admin-only' and 'DAAdministrators' not in groups) or log_config == 'disabled':
-            raise exceptions.ResourceUnauthorized(
-                username=context.username,
-                action='View Share Logs',
-                resource_uri=shareUri,
-            )
+        if (log_config == 'admin-only' and 'DAAdministrators' not in groups):
+            return False
         with context.db_engine.scoped_session() as session:
             share = ShareObjectRepository.get_share_by_uri(session, shareUri)
             ds = DatasetBaseRepository.get_dataset_by_uri(session, share.datasetUri)
